Add functionality to dynamically render skillset, skill name and skill image
