+++ src/app/skills/skills.component.css
@@ -1,6 +1,6 @@
 .Skills {
 	display: grid;
-	grid-gap: 4rem;
+	grid-gap: 3rem;
 	grid-template-columns: repeat(auto-fit, 10rem);
 	margin: 0 10rem 0 2rem;
 }
