+++ src/app/skills/skills.component.html
@@ -1,4 +1,24 @@
 <section class='Skills'>
+<table class='Skills-table' *ngFor='let skillset of skillsets'>
+		<thead>
+			<tr>
+				<th>{{skillset.category}}</th>
+			</tr>
+		</thead>
+		<tbody>
+			<tr class='tbody-row' *ngFor='let skill of skillset.skills'>
+				<td class='td-name'>{{skill.name}}</td>
+				<td class='td-image'>
+					<div class='skills-icon html' [ngStyle]="{ 'background-image': 'url(' + skill.image + ')'}"></div>
+				</td>
+			</tr>
+		</tbody>
+	</table>
+</section>
+
+
+
+<!-- <section class='Skills'>
 	<table class='Skills-table-1'>
 		<thead>
 			<tr>
@@ -220,3 +240,4 @@
 		</tbody>
 	</table>
 </section>
+ -->
