+++ src/app/skills/skills.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { SKILLSETS } from '../skills-data';
 
 @Component({
   selector: 'app-skills',
@@ -6,6 +7,7 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./skills.component.css']
 })
 export class SkillsComponent implements OnInit {
+	skillsets = SKILLSETS;
 
   constructor() { }
